pyproject.toml
[tool.poetry]
name = "superbot"
version = "0.1.0"
description = ""
authors = ["flastors"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
aiogram = "^3.10.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
motor = "^3.5.1"
tornado = "^6.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
